--- UIComponentsExample/UIComponentsExample/UIComponentsExampleApp.swift
@@ -1,17 +1,0 @@
-//
-//  UIComponentsExampleApp.swift
-//  UIComponentsExample
-//
-//  Created by Nicolai Dam on 14/02/2023.
-//
-
-import SwiftUI
-
-@main
-struct UIComponentsExampleApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-    }
-}
